Add productivity chart hover details
